Move trades sync-schema to separate file
